// apps/api/src/lib/source/imdbapidev/generated/models/imdbapi-interest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IMDbAPI
 * IMDb API for accessing movie and TV show data
 *
 * The version of the OpenAPI document: 2.7.12
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ImdbapiImage } from './imdbapi-image.js';

/**
 * Interest represents a specific interest in the IMDB API.
 * @export
 * @interface ImdbapiInterest
 */
export interface ImdbapiInterest {
    /**
     * Unique identifier for the interest.
     * @type {string}
     * @memberof ImdbapiInterest
     */
    'id'?: string;
    /**
     * The name of the interest, e.g., \"Action\", \"Action Epic\", \"Adult Animation\", etc.
     * @type {string}
     * @memberof ImdbapiInterest
     */
    'name'?: string;
    /**
     * 
     * @type {ImdbapiImage}
     * @memberof ImdbapiInterest
     */
    'primaryImage'?: ImdbapiImage;
    /**
     * A brief description of the interest, which can include details about the genre or type.
     * @type {string}
     * @memberof ImdbapiInterest
     */
    'description'?: string;
    /**
     * Indicates whether the interest is a subgenre of another genre.
     * @type {boolean}
     * @memberof ImdbapiInterest
     */
    'isSubgenre'?: boolean;
    /**
     * A list of similar interests that are related to this interest.
     * @type {Array<ImdbapiInterest>}
     * @memberof ImdbapiInterest
     */
    'similarInterests'?: Array<ImdbapiInterest>;
}